+++ src/core_modules/capture-core/components/Pages/EnrollmentEditEvent/EnrollmentEditEventPage.types.js
@@ -2,6 +2,7 @@
 import type { ProgramStage } from '../../../metaData';
 import type { WidgetEffects, HideWidgets } from '../common/EnrollmentOverviewDomain';
 import type { UserFormField } from '../../FormFields/UserField';
+import type { LinkedRecordClick } from '../../WidgetsRelationship/WidgetTrackedEntityRelationship';
 
 export type PlainProps = {|
     programStage: ?ProgramStage,
@@ -10,6 +11,7 @@ export type PlainProps = {|
     teiId: string,
     enrollmentId: string,
     programId: string,
+    trackedEntityTypeId: string,
     mode: string,
     orgUnitId: string,
     trackedEntityName: string,
@@ -20,6 +22,7 @@ export type PlainProps = {|
     onDelete: () => void,
     onAddNew: () => void,
     onGoBack: () => void,
+    onLinkedRecordClick: LinkedRecordClick,
     onEnrollmentError: (message: string) => void,
     onEnrollmentSuccess: () => void,
     onCancelEditEvent: (isScheduled: boolean) => void,