+++ src/core_modules/capture-core/components/WidgetsRelationship/common/RelationshipsWidget/relationshipsWidget.types.js
@@ -1,0 +1,19 @@
+// @flow
+import type { Node } from 'React';
+import type { InputRelationshipData, RelationshipTypes } from '../Types';
+import type { LinkedRecordClick } from './types';
+
+export type Props = $ReadOnly<{|
+    title: string,
+    relationships?: Array<InputRelationshipData>,
+    cachedRelationshipTypes?: RelationshipTypes,
+    isLoading: boolean,
+    sourceId: string,
+    onLinkedRecordClick: LinkedRecordClick,
+    children: (relationshipTypes: RelationshipTypes) => Node,
+|}>;
+
+export type StyledProps = $ReadOnly<{|
+    ...Props,
+    ...CssClasses,
+|}>;